factor out scheduler initialization in initializeTriggerObjects
